refactor: remove background booting
